// app/patients/model/patient.ts
export class Patient {
    Id: number;
    School: string;
    NameFather: string;
    LastNamesFather: string;
    WorkFather: string;
    NameMother: string;
    LastNamesMother: string;
    WorkMother: string;
    ReasonForConsultation: string;
    CurrentProblem: string;
    personalAndFamilyHistory:any;
    vitalSigns:any;
    stomatognathicSystem: any;
    NotePersonalAndFamilyRecord: string;
    NoteReviewSystemStomatognathic: string;
    PersonId: number;
    Person: any;
    ClinicId: number;
    Patients_Catalog: any;
    Patients_Catalog_By_Type: any;
    Age: number;
    PatientFromPartner: boolean;
}